finish generateXlsx and updateLocalesFromXlsx

+++ packages/core/src/utils/help.ts
@@ -1,6 +1,20 @@
 import path from 'path'
 import fs from 'fs'
 import { ESLint } from 'eslint'
+import consola from 'consola'
+
+const createFileName = (fileName = 'fileName') => {
+  const data = new Date()
+  const name = `${fileName}(${
+    data.getMonth() + 1
+  }-${data.getDate()} ${data.getHours()}:${
+    data.getMinutes() <= 9 ? `0${data.getMinutes()}` : data.getMinutes()
+  }:${
+    data.getSeconds() <= 9 ? `0${data.getSeconds()}` : data.getSeconds()
+  })}`
+
+  return name
+}
 
 const lexicalComparator = (a: string, b: string) => {
   a = a.toString()
@@ -46,19 +60,22 @@ const writeToJsonFile = async (
   obj: object,
 ) => {
   const jsonPath = path.join(writeToPath, `${name}.json`)
-  fs.writeFileSync(jsonPath, JSON.stringify(sortObjectKey(obj)), 'utf8')
-  const eslint = new ESLint({
-    fix: true,
-    useEslintrc: false,
-    baseConfig: require(path.join(path.resolve(), '.eslintrc.js')),
-  })
-  const results = await eslint.lintFiles(jsonPath)
+  fs.writeFileSync(jsonPath, JSON.stringify(sortObjectKey(obj), undefined, 4), 'utf8')
 
-  if (results[0].output)
-    fs.writeFileSync(jsonPath, results[0].output, 'utf8')
+  const eslintrcPath = path.join(path.resolve(), '.eslintrc.js')
+  if (fs.existsSync(eslintrcPath)) {
+    const eslint = new ESLint({
+      fix: true,
+      useEslintrc: true,
+      baseConfig: (await import(eslintrcPath)).default,
+    })
+    const results = await eslint.lintFiles(jsonPath)
 
-  else
-    console.log('eslint Error!', results[0])
+    if (results[0].output)
+      fs.writeFileSync(jsonPath, results[0].output, 'utf8')
+    else
+      consola.error('eslint Error!', results[0])
+  }
 }
 
 const getValueByKey = (obj: any, key: string) => {
@@ -72,7 +89,6 @@ const getValueByKey = (obj: any, key: string) => {
 
     if (path && temp[path] !== undefined)
       temp = temp[path]
-
     else
       return undefined
   }
@@ -82,10 +98,8 @@ const getValueByKey = (obj: any, key: string) => {
 const setValueByKey = (obj: any, key: string, val: string) => {
   const paths = key.split('.').filter(x => x !== '')
 
-  if (paths.length <= 1) {
-    console.error(`can not set ${val} to the ${key}`)
+  if (paths.length === 0)
     return
-  }
 
   let temp = obj
   while (paths.length > 1) {
@@ -97,7 +111,7 @@ const setValueByKey = (obj: any, key: string, val: string) => {
 
       temp = temp[path]
       if (typeof temp !== 'object') {
-        console.error(`can not set ${val} to the ${key}`)
+        consola.error(`can not set ${val} to the ${key}`)
         return
       }
     }
@@ -112,9 +126,10 @@ const getKeys = (obj: any) => {
     for (const k of Object.keys(obj)) {
       if (typeof obj[k] === 'string')
         res.push(_before === '' ? k : `${_before}.${k}`)
-
-      else
+      else if (typeof obj[k] === 'object' && !(Array.isArray(obj[k])))
         res.push(..._getKeys(obj[k], _before === '' ? k : `${_before}.${k}`))
+      else
+        consola.error(`the i18n JSON value only support string (${k}: ${obj[k]})`)
     }
     return res
   }
@@ -128,4 +143,5 @@ export {
   getValueByKey,
   setValueByKey,
   getKeys,
+  createFileName,
 }
